requirements.txt
streamlit
tavily
google-generativeai